added another hook and clean up
