+++ frontend/src/Socket.js
@@ -1,5 +1,6 @@
 import socketIOClient from 'socket.io-client';
 
-const socket = new socketIOClient('http://localhost:3000/');
+// const socket = new socketIOClient('mafia-hands-backend-cd.herokuapp.com');
+const socket = new socketIOClient('localhost:3000');
 
 export default socket;
